--- src/main/java/tutorlink/appstate/AppState.java
@@ -1,14 +1,0 @@
-package tutorlink.appstate;
-
-import tutorlink.lists.ComponentList;
-import tutorlink.lists.GradeList;
-import tutorlink.lists.StudentList;
-
-public class AppState {
-    public StudentList students = new StudentList();
-    public GradeList grades = new GradeList();
-    public ComponentList components = new ComponentList();
-
-    public AppState() {
-    }
-}
