+++ src/test/java/tutorlink/command/DeleteComponentCommandTest.java
@@ -3,14 +3,15 @@ package tutorlink.command;
 import java.util.HashMap;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import tutorlink.appstate.AppState;
-import tutorlink.commons.Commons;
-import tutorlink.component.Assignment;
-import tutorlink.component.ClassParticipation;
-import tutorlink.component.Exam;
-import tutorlink.exceptions.ComponentNotFoundException;
-import tutorlink.exceptions.IllegalValueException;
-import tutorlink.result.CommandResult;
+import tutorlink.logic.command.DeleteComponentCommand;
+import tutorlink.model.AppState;
+import tutorlink.model.Commons;
+import tutorlink.model.component.Assignment;
+import tutorlink.model.component.ClassParticipation;
+import tutorlink.model.component.Exam;
+import tutorlink.model.exceptions.ComponentNotFoundException;
+import tutorlink.model.exceptions.IllegalValueException;
+import tutorlink.logic.result.CommandResult;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
