+++ src/test/java/tutorlink/command/FindStudentCommandTest.java
@@ -4,11 +4,13 @@ package tutorlink.command;
 import java.util.HashMap;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import tutorlink.appstate.AppState;
-import tutorlink.commons.Commons;
-import tutorlink.exceptions.IllegalValueException;
-import tutorlink.exceptions.StudentNotFoundException;
-import tutorlink.result.CommandResult;
+import tutorlink.logic.command.AddStudentCommand;
+import tutorlink.logic.command.FindStudentCommand;
+import tutorlink.model.AppState;
+import tutorlink.model.Commons;
+import tutorlink.model.exceptions.IllegalValueException;
+import tutorlink.model.exceptions.StudentNotFoundException;
+import tutorlink.logic.result.CommandResult;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.fail;
