+++ src/test/java/tutorlink/lists/GradeListTest.java
@@ -3,13 +3,14 @@ package tutorlink.lists;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import tutorlink.component.Assignment;
-import tutorlink.component.Exam;
-import tutorlink.exceptions.DuplicateGradeException;
-import tutorlink.exceptions.GradeNotFoundException;
-import tutorlink.grade.Grade;
-import tutorlink.student.Student;
-import tutorlink.component.Component;
+import tutorlink.model.component.Assignment;
+import tutorlink.model.component.Exam;
+import tutorlink.model.exceptions.DuplicateGradeException;
+import tutorlink.model.exceptions.GradeNotFoundException;
+import tutorlink.model.grade.Grade;
+import tutorlink.model.lists.GradeList;
+import tutorlink.model.student.Student;
+import tutorlink.model.component.Component;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
